+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Features
 
 1. [#302](https://github.com/influxdata/influxdb-client-js/pull/302): Expose response headers in all APIs.
-1. [#302](https://github.com/influxdata/influxdb-client-js/pull/302): Add `sessionAuth.js` example.
+1. [#302](https://github.com/influxdata/influxdb-client-js/pull/302): Add `tokens.js` example.
 
 ## 1.10.0 [2021-01-29]